Refactor: Add user.repository.mock to reuse in tests
