+++ src/users/users.controller.spec.ts
@@ -1,25 +1,9 @@
 import { Test, TestingModule } from '@nestjs/testing';
 import { getRepositoryToken } from '@nestjs/typeorm';
+import { MockUsersRepository } from '../../test/mock/user.repository.mock';
 import { User } from './user.entity';
 import { UsersController } from './users.controller';
 import { UsersService } from './users.service';
-
-const mockUser = {
-	id: '40118a2e-dc89-46de-bbb0-46b8ac0c6846',
-	createdAt: '2022-07-01T01:43:24.252Z',
-	updatedAt: '2022-07-01T01:43:24.252Z',
-	deletedAt: null,
-	email: 'testtest@email.com',
-	name: 'test chung',
-	password: '12345',
-};
-class MockUsersRepository {
-	save = jest.fn().mockResolvedValue(mockUser);
-
-	async find() {
-		return mockUser;
-	}
-}
 
 describe('UsersController', () => {
 	let controller: UsersController;
